refactor(InputNumbers): InputNumberFactory 제거 & InputNumbers로 검증 메서드 병합에 따른 생성자 수정
InputNumberFactory의 역할을 InputNumbers가 충분히 수행할 수 있을 것 같다고 판단했습니다. 처음에는 생성을 위임하고자 만들었었으나, 필요 없는 클래스라고 판단되어 삭제하였습니다.
이에 따라, InputNumbers로 validation 로직들을 옮겼고, 본래 List<SingleNumber>를 파라미터로 받아 생성하던 생성자를 삭제하고, String inputValue를 받아 생성하는 생성자로 변경했습니다.
valdation은 객체가 생성이 될 때 수행되도록 하였습니다.
