+++ src/test/java/baseball/domain/inputnumber/InputNumbersTest.java
@@ -1,19 +1,36 @@
 package baseball.domain.inputnumber;
 
-import baseball.domain.inputnumber.InputNumberFactory;
-import baseball.domain.inputnumber.InputNumbers;
 import baseball.domain.SingleNumber;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class InputNumbersTest {
 
+    @ParameterizedTest(name = "중복_숫자를_입력하면_예외를_던진다")
+    @ValueSource(strings = {"112", "111"})
+    void 중복_숫자를_입력하면_예외를_던진다(String inputValue) {
+        assertThatThrownBy(() -> new InputNumbers(inputValue))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessageContaining("중복 숫자는 입력할 수 없습니다.");
+    }
+
+    @ParameterizedTest(name = "세_자리가_아닌_문자열을_입력받으면_예외를_던진다")
+    @ValueSource(strings = {"1234", "12"})
+    void 세_자리가_아닌_문자열을_입력받으면_예외를_던진다(String inputValue) {
+        assertThatThrownBy(() -> new InputNumbers(inputValue))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessageContaining("세자리의 숫자를 입력해 주세요.");
+    }
+
     @Test
     void ball이_3일_경우_3을_리턴한다() {
-        InputNumbers inputNumbers = InputNumberFactory.newInstance("123");
+        InputNumbers inputNumbers = new InputNumbers("123");
         List<SingleNumber> randomNumbers = List.of(
                 new SingleNumber(0, 3),
                 new SingleNumber(1, 1),
@@ -25,7 +42,7 @@ class InputNumbersTest {
 
     @Test
     void ball이_0일_경우_0을_리턴한다() {
-        InputNumbers inputNumbers = InputNumberFactory.newInstance("123");
+        InputNumbers inputNumbers = new InputNumbers("123");
         List<SingleNumber> randomNumbers = List.of(
                 new SingleNumber(0, 1),
                 new SingleNumber(1, 2),
@@ -37,7 +54,7 @@ class InputNumbersTest {
 
     @Test
     void strike가_3일_경우_3을_리턴한다() {
-        InputNumbers inputNumbers = InputNumberFactory.newInstance("123");
+        InputNumbers inputNumbers = new InputNumbers("123");
         List<SingleNumber> randomNumbers = List.of(
                 new SingleNumber(0, 1),
                 new SingleNumber(1, 2),
@@ -49,7 +66,7 @@ class InputNumbersTest {
 
     @Test
     void strike가_0일_경우_0을_리턴한다() {
-        InputNumbers inputNumbers = InputNumberFactory.newInstance("123");
+        InputNumbers inputNumbers = new InputNumbers("123");
         List<SingleNumber> randomNumbers = List.of(
                 new SingleNumber(0, 3),
                 new SingleNumber(1, 1),
@@ -58,4 +75,5 @@ class InputNumbersTest {
 
         assertThat(inputNumbers.strikeCount(randomNumbers)).isEqualTo(0);
     }
+
 }
